+++ packages/router/router.ts
@@ -735,17 +735,15 @@ export function createRouter(init: RouterInit): Router {
   ): void {
     // Deduce if we're in a loading/actionReload state:
     // - We have committed actionData in the store
-    // - The current navigation was a submission
+    // - The current navigation was a mutation submission
     // - We're past the submitting state and into the loading state
-    // - The location we've finished loading is different from the submission
-    //   location, indicating we redirected from the action (avoids false
-    //   positives for loading/submissionRedirect when actionData returned
-    //   on a prior submission)
+    // - The location being loaded is not the result of a redirect
     let isActionReload =
       state.actionData != null &&
       state.navigation.formMethod != null &&
+      isMutationMethod(state.navigation.formMethod) &&
       state.navigation.state === "loading" &&
-      state.navigation.formAction?.split("?")[0] === location.pathname;
+      location.state?._isRedirect !== true;
 
     let actionData: RouteData | null;
     if (newState.actionData) {
@@ -1092,7 +1090,7 @@ export function createRouter(init: RouterInit): Router {
         replace =
           result.location === state.location.pathname + state.location.search;
       }
-      await startRedirectNavigation(state, result, replace);
+      await startRedirectNavigation(state, result, { submission, replace });
       return { shortCircuited: true };
     }
 
@@ -1152,10 +1150,26 @@ export function createRouter(init: RouterInit): Router {
       loadingNavigation = navigation;
     }
 
+    // If this was a redirect from an action we don't have a "submission" but
+    // we have it on the loading navigation so use that if available
+    let activeSubmission = submission
+      ? submission
+      : loadingNavigation.formMethod &&
+        loadingNavigation.formAction &&
+        loadingNavigation.formData &&
+        loadingNavigation.formEncType
+      ? {
+          formMethod: loadingNavigation.formMethod,
+          formAction: loadingNavigation.formAction,
+          formData: loadingNavigation.formData,
+          formEncType: loadingNavigation.formEncType,
+        }
+      : undefined;
+
     let [matchesToLoad, revalidatingFetchers] = getMatchesToLoad(
       state,
       matches,
-      submission,
+      activeSubmission,
       location,
       isRevalidationRequired,
       cancelledDeferredRoutes,
@@ -1244,7 +1258,7 @@ export function createRouter(init: RouterInit): Router {
     // If any loaders returned a redirect Response, start a new REPLACE navigation
     let redirect = findRedirect(results);
     if (redirect) {
-      await startRedirectNavigation(state, redirect, replace);
+      await startRedirectNavigation(state, redirect, { replace });
       return { shortCircuited: true };
     }
 
@@ -1401,7 +1415,9 @@ export function createRouter(init: RouterInit): Router {
       state.fetchers.set(key, loadingFetcher);
       updateState({ fetchers: new Map(state.fetchers) });
 
-      return startRedirectNavigation(state, actionResult, false, true);
+      return startRedirectNavigation(state, actionResult, {
+        isFetchActionRedirect: true,
+      });
     }
 
     // Process any non-redirect errors thrown
@@ -1673,8 +1689,15 @@ export function createRouter(init: RouterInit): Router {
   async function startRedirectNavigation(
     state: RouterState,
     redirect: RedirectResult,
-    replace?: boolean,
-    isFetchActionRedirect?: boolean
+    {
+      submission,
+      replace,
+      isFetchActionRedirect,
+    }: {
+      submission?: Submission;
+      replace?: boolean;
+      isFetchActionRedirect?: boolean;
+    } = {}
   ) {
     if (redirect.revalidate) {
       isRevalidationRequired = true;
@@ -1714,24 +1737,30 @@ export function createRouter(init: RouterInit): Router {
     let redirectHistoryAction =
       replace === true ? HistoryAction.Replace : HistoryAction.Push;
 
+    // Use the incoming submission if provided, fallback on the active one in
+    // state.navigation
     let { formMethod, formAction, formEncType, formData } = state.navigation;
+    if (!submission && formMethod && formAction && formData && formEncType) {
+      submission = {
+        formMethod,
+        formAction,
+        formEncType,
+        formData,
+      };
+    }
 
     // If this was a 307/308 submission we want to preserve the HTTP method and
     // re-submit the GET/POST/PUT/PATCH/DELETE as a submission navigation to the
     // redirected location
     if (
       redirectPreserveMethodStatusCodes.has(redirect.status) &&
-      formMethod &&
-      isMutationMethod(formMethod) &&
-      formEncType &&
-      formData
+      submission &&
+      isMutationMethod(submission.formMethod)
     ) {
       await startNavigation(redirectHistoryAction, redirectLocation, {
         submission: {
-          formMethod,
+          ...submission,
           formAction: redirect.location,
-          formEncType,
-          formData,
         },
       });
     } else {
@@ -1741,10 +1770,10 @@ export function createRouter(init: RouterInit): Router {
         overrideNavigation: {
           state: "loading",
           location: redirectLocation,
-          formMethod: formMethod || undefined,
-          formAction: formAction || undefined,
-          formEncType: formEncType || undefined,
-          formData: formData || undefined,
+          formMethod: submission ? submission.formMethod : undefined,
+          formAction: submission ? submission.formAction : undefined,
+          formEncType: submission ? submission.formEncType : undefined,
+          formData: submission ? submission.formData : undefined,
         },
       });
     }