+++ packages/router/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@remix-run/router`
 
+## 1.2.1
+
+### Patch Changes
+
+- Include submission info in `shouldRevalidate` on action redirects ([#9777](https://github.com/remix-run/react-router/pull/9777), [#9782](https://github.com/remix-run/react-router/pull/9782))
+- Reset `actionData` on action redirect to current location ([#9772](https://github.com/remix-run/react-router/pull/9772))
+
 ## 1.2.0
 
 ### Minor Changes